refactor resource detail header + add basic like btn
