+++ frontend/src/components/icons.jsx
@@ -1295,3 +1295,42 @@ export const PinForm = () => (
     />
   </svg>
 )
+
+export const Like = () => {
+  return (
+    <svg
+      width="16"
+      height="14"
+      viewBox="0 0 16 14"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M1.52212 5.78394H3.99692V12.3736H1.52212C1.45355 12.3736 1.38566 12.3601 1.32231 12.3338C1.25897 12.3076 1.20141 12.2691 1.15293 12.2207C1.10444 12.1722 1.06598 12.1146 1.03974 12.0513C1.0135 11.9879 1 11.92 1 11.8515V6.30605C1 6.16758 1.05501 6.03478 1.15293 5.93686C1.25084 5.83894 1.38364 5.78394 1.52212 5.78394Z"
+        fill="white"
+      />
+      <path
+        d="M1.52212 5.78394H3.99692V12.3736H1.52212C1.45355 12.3736 1.38566 12.3601 1.32231 12.3338C1.25897 12.3076 1.20141 12.2691 1.15293 12.2207C1.10444 12.1722 1.06598 12.1146 1.03974 12.0513C1.0135 11.9879 1 11.92 1 11.8515V6.30605C1 6.16758 1.05501 6.03478 1.15293 5.93686C1.25084 5.83894 1.38364 5.78394 1.52212 5.78394Z"
+        stroke="#4147D5"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+      />
+      <path
+        d="M3.99683 5.78401L6.64953 1.5439C6.75232 1.3762 6.89684 1.23801 7.06897 1.14281C7.2411 1.04761 7.43497 0.99866 7.63166 1.00073C7.78639 0.995239 7.94066 1.02081 8.08535 1.07593C8.23004 1.13105 8.36222 1.21461 8.47408 1.32166C8.58594 1.42871 8.67522 1.5571 8.73664 1.69923C8.79806 1.84136 8.83038 1.99435 8.83169 2.14918V5.18821H13.4171C13.5865 5.19375 13.7528 5.23509 13.9051 5.30952C14.0574 5.38394 14.1922 5.48976 14.3007 5.62001C14.4092 5.75026 14.4889 5.902 14.5345 6.06525C14.5802 6.2285 14.5907 6.39956 14.5655 6.56719L13.7308 11.9568C13.6949 12.2446 13.5552 12.5093 13.3379 12.7014C13.1206 12.8935 12.8407 12.9997 12.5507 13H5.69898C5.37265 13.001 5.05159 12.9263 4.76 12.781L4.00735 12.4052"
+        fill="#D7E0FF"
+      />
+      <path
+        d="M3.99683 5.78401L6.64953 1.5439C6.75232 1.3762 6.89684 1.23801 7.06897 1.14281C7.2411 1.04761 7.43497 0.99866 7.63166 1.00073C7.78631 0.995241 7.94049 1.02078 8.08511 1.07583C8.22973 1.13089 8.36186 1.21435 8.47371 1.32129C8.58555 1.42823 8.67485 1.55649 8.73634 1.6985C8.79782 1.84051 8.83024 1.99339 8.83169 2.14813V5.18715H13.4171C13.5865 5.19253 13.753 5.23377 13.9053 5.30814C14.0577 5.38251 14.1926 5.48833 14.3011 5.61862C14.4096 5.74891 14.4893 5.90072 14.5349 6.06404C14.5804 6.22736 14.5909 6.39849 14.5655 6.56614L13.7308 11.9547C13.6951 12.2427 13.5555 12.5077 13.3382 12.7C13.1209 12.8923 12.8409 12.9986 12.5507 12.9989H5.69898C5.37265 13 5.05159 12.9253 4.76 12.78L4.00735 12.4042"
+        stroke="#4147D5"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+      />
+      <path
+        d="M3.99683 5.78394V12.3736"
+        stroke="#4147D5"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+      />
+    </svg>
+  )
+}
